Created Scrollable Paged Index File
Added init files
Refactored view control; moved rendering to views.py, finalized static
proxy views, added preliminary ajax html file loading.
