+++ modules/app/static/js/index_script.js
@@ -1,0 +1,6 @@
+$(document).ready(function(){
+    $("#extract").load("../templates/extract.html");
+    $("#pca").load("../templates/pca.html");
+    $("#model").load("../templates/model.html");
+    $("#viz").load("../templates/viz.html");
+});
