+++ modules/app/templates/index.html
@@ -3,9 +3,11 @@
     <head>
         <meta charset="UTF-8">
         <title>MoSeq2-App Home</title>
+        <link rel="icon" href="data:;base64,iVBORw0KGgo="> <!-- suppress GET error -->
         <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
         <link rel="stylesheet" href="/static/css/index_style.css">
         <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
         <script src="/static/js/index_script.js"></script>
         <meta name="description" content="MoSeq2 Application Home Page">
         <meta name="author" content="Ayman Zeine">
@@ -15,10 +17,45 @@
 
     <body>
 
-        <div class="align-middle">
+        <div class="topnav">
+            <ul id="horizontal-list">
+                <li><a class="active" href="#home">Home</a></li>
+                <li><a href="#extract">Extract</a></li>
+                <li><a href="#pca">PCA</a></li>
+                <li><a href="#model">Model</a></li>
+                <li><a href="#viz">Viz</a></li>
+                <li><a href="#contact">Contact</a></li>
+                <li><a href="#contribute">Contribute</a></li>
+            </ul>
+        </div>
+        <div id="main">
+            <div id="home">
+                <h1>Welcome to the MoSeq2 Application Home Page</h1>
+            </div>
 
-            <h1>Welcome to the MoSeq2 Application Home Page</h1>
+            <div id="extract">
 
+            </div>
+
+            <div id="pca">
+
+            </div>
+
+            <div id="model">
+
+            </div>
+
+            <div id="viz">
+
+            </div>
+
+            <div id="contact">
+
+            </div>
+
+            <div id="contribute">
+
+            </div>
         </div>
 
     </body>
