+++ modules/app/templates/model.html
@@ -3,6 +3,7 @@
     <head>
         <meta charset="UTF-8">
         <title>MoSeq2-Model Operation</title>
+        <link rel="icon" href="data:;base64,iVBORw0KGgo="> <!-- suppress GET error -->
         <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
         <link rel="stylesheet" href="/static/css/model_style.css">
         <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
