+++ uwsgi.ini
@@ -1,0 +1,4 @@
+[uwsgi]
+module = main
+callable = app
+master = true
